FastFFT: part 15
Fixes resampling (downsampling) during projection.

Templates GPU projection and allows reading CTF image from a pointer (as before) or from a texture. The change is to both speed up the reads (tex = read only cache)
but also to allow lerp of the CTF during LERP downsample projection. This greatly improves the quality of the projections compared to
using a CTFImage created at the cropped size.