FastFFT: part 17
With the replicative padding introduced in the previous commit, some images that had artificats or tapering at their edges were producing strong central cross artifacts.
This commit adds the command line option to apply a mask to the central cross when resizing the image. Without it, the average mip values have a strong depenency on the amount of padding, which
is obviously not acceptable. With the mask, the small variance depending on padding matches the use of Gaussian noise padding, but with a systematically lower SNR. Keeping this for future
trouble-shooting, but I'm going to leave a note to prefer Gaussian padding.

Changes  Image::ReturnLinearInterpolated2D(..) to take two const floats rather than two float refs.
    - The input args are never modified by the function, but being refs requires them to be lvalues, so something like Image::ReturnLinearInterpolated2D(my_other_func() -> returns float, 2.0f) would not work.

Changes TemplateMatchingDataSizer::ResizeImage_preSearch to:
    - mask the central cross when resizing. In some images, it seems a small taper to the values near the edge produces strong central cross artifacts, especially with replicative padding.
        - The value is fixed at a width of 1.0, this may not be optimal
        - There is no option to turn it off.

Lowers the threshold to consider mipping to -0.5 (from 1.0f in the current branch, and what is 0.f in the master branch)
    - Some search images were producing mips with small "holes" where no values were found > 1.0

Changes behavior of outlier removal to only happen once in TemplateMatchingDataSizer::PreProcessInputImage, where previously it also was being done after resizing. I did not observe any negative effects, but
this was a deviation from the design I hadn't intended, so I "fixed" it.